Move mangata files to mangata folder
