+++ package.json
@@ -1,9 +1,9 @@
 {
   "scripts": {
-    "mangata-rococo": "dotenv -e .env babel-node src/mangata-rococo.js",
-    "mangata-kusama": "dotenv -e .env babel-node src/mangata-kusama.js",
-    "mangata": "dotenv -e .env babel-node src/mangata.js",
-    "setup-mangata": "dotenv -e .env babel-node src/setup-mangata.js",
+    "mangata-rococo": "dotenv -e .env babel-node src/mangata/mangata-rococo.js",
+    "mangata-kusama": "dotenv -e .env babel-node src/mangata/mangata-kusama.js",
+    "mangata": "dotenv -e .env babel-node src/mangata/mangata.js",
+    "setup-mangata": "dotenv -e .env babel-node src/mangata/setup-mangata.js",
     "shibuya": "dotenv -e .env babel-node src/shibuya.js",
     "rocstar": "dotenv -e .env babel-node src/rocstar.js",
     "utils": "dotenv -e .env babel-node src/utils.js",
