+++ src/common/account.js
@@ -1,9 +1,8 @@
 import _ from 'lodash';
 import util from 'util';
 import Keyring from '@polkadot/keyring';
+
 import { getDecimalBN } from './utils';
-import turingHelper from './turingHelper';
-import mangataHelper from './mangataHelper';
 
 // Create a keyring instance
 const keyring = new Keyring({ type: 'sr25519' });
